+++ client/views/directory/hooks.js
